Changes probe result status meaning (#355)

+++ core/src/main/java/io/jenkins/pluginhealth/scoring/service/ScoreService.java
@@ -86,7 +86,7 @@ public class ScoreService {
         return repository.deleteOldScoreFromPlugin();
     }
 
-    public record ScoreStatistics(int average, int minimum, int maximum, int firstQuartile, int median,
+    public record ScoreStatistics(double average, int minimum, int maximum, int firstQuartile, int median,
                                   int thirdQuartile) {
     }
 }
